+++ lending/src/sdk/models/operations/getcategorizedbalancesheetstatement.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class GetCategorizedBalanceSheetStatementRequest extends SpeakeasyBase {
@@ -40,7 +40,7 @@ export class GetCategorizedBalanceSheetStatementResponse extends SpeakeasyBase {
     enhancedFinancialReport?: shared.EnhancedFinancialReport;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
@@ -55,5 +55,5 @@ export class GetCategorizedBalanceSheetStatementResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }