+++ lending/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v3.3.0] lending
 ### Releases
 - [NPM v3.3.0] https://www.npmjs.com/package/@codat/lending/v/3.3.0 - lending
+
+## 2023-11-27 01:29:32
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0] lending
+### Releases
+- [NPM v4.0.0] https://www.npmjs.com/package/@codat/lending/v/4.0.0 - lending